The in-memory file systems (memFS: and memLZF:) did not support files larger than 2 GB

+++ h2/src/main/org/h2/store/fs/FilePathMem.java
@@ -283,7 +283,7 @@ class FileMem extends FileBase {
 
     @Override
     public FileChannel position(long newPos) {
-        this.pos = (int) newPos;
+        this.pos = newPos;
         return this;
     }
